# Levara/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Levara HNSW + BM25 + Neo4j vector DB HTTP API.\nSee CLAUDE.md for architecture; pkg/mcp for the MCP alternate transport.",
        "title": "Levara API",
        "contact": {},
        "version": "v1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/datasets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns datasets owned by the caller plus any explicitly shared. Superusers see all rows.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasets"
                ],
                "summary": "List datasets visible to the caller",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/http.DatasetDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasets"
                ],
                "summary": "Create a new dataset",
                "parameters": [
                    {
                        "description": "Dataset name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/http.DatasetDTO"
                        }
                    },
                    "400": {
                        "description": "name required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/datasets/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasets"
                ],
                "summary": "Delete a dataset (idempotent)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dataset UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.DatasetDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "record_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003cjwt\u003e\" from POST /auth/login or an API key from POST /auth/keys.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}